+++ GaluhadPlugins/Pets/Strings/English/Pets.lua
@@ -555,87 +555,89 @@ _PETSTRINGS[467] = {[1]="Crimson Deep-claw Whelp";[2]="Available from Hobbit Pre
 _PETSTRINGS[468] = {[1]="Sapphire Drakeling";[2]="Limited time availability: Legendary Item Reward Track Season 16."}
 _PETSTRINGS[469] = {[1]="Helpful Deep-claw Whelp";[2]="Limited time availability: Legendary Item Reward Track Season 16."}
 
-_PETSTRINGS[470] = {[1]="Gold Zizanu";[2]=""}
-_PETSTRINGS[471] = {[1]="Dusty Worm";[2]=""}
-_PETSTRINGS[472] = {[1]="Well-laden Mûmak Calf";[2]=""}
-_PETSTRINGS[473] = {[1]="Redstone Mûmak Calf";[2]=""}
+_PETSTRINGS[470] = {[1]="Gold Zizanu";[2]="Available through barter from the City of Zajâna Rewards Quartermaster in Zajâna in Kighân."}
+_PETSTRINGS[471] = {[1]="Dusty Worm";[2]="Available through barter from the Kintai of Sul Madásh Rewards Quartermaster in Sul Madásh in Adagím."}
+_PETSTRINGS[472] = {[1]="Well-laden Mûmak Calf";[2]="Available through barter from the Temámir of Jiret-menêsh Rewards Quartermaster in Jiret-menêsh in An Shêru."}
+_PETSTRINGS[473] = {[1]="Redstone Mûmak Calf";[2]="Available through barter from Hamât Renewed Rewards Quartermasters in Idagâl and Pahar Hatokáli."}
 _PETSTRINGS[474] = {[1]="Vibrant Spotted Boglurker";[2]="Limited time availability: Legendary Item Reward Track Season 17."}
 _PETSTRINGS[475] = {[1]="Rose-coloured Mûmak Calf";[2]="Limited time availability: Legendary Item Reward Track Season 17."}
-_PETSTRINGS[476] = {[1]="Marine Scorpion";[2]=""}
+_PETSTRINGS[476] = {[1]="Death-light Scorpion";[2]="Available through barter with the Hunter's Guild of Mûr Ghala"}
 _PETSTRINGS[477] = {[1]="Daffodil Celebration Kid";[2]="Available through barter during the Spring Festival."}
 _PETSTRINGS[478] = {[1]="Tawny Aurochs";[2]="Available through barter during the Spring Festival."}
+
+_PETSTRINGS[479] = {[1]="Quintessential Fawn";[2]="Available through barter during the Anniversary Festival."}
+_PETSTRINGS[480] = {[1]="Brown Jerboa";[2]="Available through barter during the Anniversary Festival."}
+_PETSTRINGS[481] = {[1]="Dusky Parrot";[2]="Available through barter during the Anniversary Festival."}
+_PETSTRINGS[482] = {[1]="Ginger Mûrai Herding Dog";[2]="Available through barter during the Anniversary Festival."}
+_PETSTRINGS[483] = {[1]="Overgrown Boglurker";[2]="Limited time random monster drop during the Anniversary Festival.";}
+_PETSTRINGS[484] = {[1]="Sunstone Drakeling";[2]="Limited time availability from 'A Special Celebration - Year Nineteen'.";}
+_PETSTRINGS[485] = {[1]="Brown Oryx";[2]="Limited time availability for Figments of Splendour and in Traveller's and Adventurer's Lootboxes.";}
+_PETSTRINGS[486] = {[1]="Young Brown Oryx";[2]="Limited time availability for Figments of Splendour and in Traveller's and Adventurer's Lootboxes."}
+_PETSTRINGS[487] = {[1]="Humming Sunbird";[2]="Limited time availability in Traveller's and Adventurer's Lootboxes."}
+_PETSTRINGS[488] = {[1]="Cobalt Grodbog Protector";[2]="Available through barter by participating in Delvings."}
+_PETSTRINGS[489] = {[1]="Obsidian Deep-claw Whelp";[2]="Available through barter by participating in Delvings."}
+_PETSTRINGS[490] = {[1]="Tan Wolf-dog";[2]="Available by aiding Shâra Mizâdi in her glorious hunt in the Hatokáli Fells in Mûr Ghala."}
+_PETSTRINGS[491] = {[1]="Young Tawny Oryx";[2]="Available through barter from Hamât Renewed Rewards Quartermasters in Idagâl and Pahar Hatokáli."}
+_PETSTRINGS[492] = {[1]="Copper Scorpion";[2]="Available through barter from Hamât Renewed Rewards Quartermasters in Idagâl and Pahar Hatokáli."}
+_PETSTRINGS[493] = {[1]="Tawny Oryx";[2]="Available through barter from Hamât Renewed Rewards Quartermasters in Idagâl and Pahar Hatokáli."}
 
 _PETSTRINGS[1000] = {[1]="Stripey Bat";[2]=""}
 _PETSTRINGS[1001] = {[1]="Tan Bat";[2]=""}
 _PETSTRINGS[1002] = {[1]="Black-wing Kite";[2]=""}
 _PETSTRINGS[1003] = {[1]="Dusk-wing Kite";[2]=""}
-_PETSTRINGS[1004] = {[1]="Tan Wolf-dog";[2]=""}
-_PETSTRINGS[1005] = {[1]="Cream Corgi";[2]=""}
-_PETSTRINGS[1006] = {[1]="Library Corgi";[2]=""}
-_PETSTRINGS[1007] = {[1]="Winter's Wind";[2]=""}
-_PETSTRINGS[1008] = {[1]="Brown Jerboa";[2]=""}
-_PETSTRINGS[1009] = {[1]="Turquoise Snake";[2]=""}
-_PETSTRINGS[1010] = {[1]="Leafy Green Snake";[2]=""}
-_PETSTRINGS[1011] = {[1]="Fiery Desert Snake";[2]=""}
-_PETSTRINGS[1012] = {[1]="Chestnut Sand Worm";[2]=""}
-_PETSTRINGS[1013] = {[1]="Spotted Hamster";[2]=""}
-_PETSTRINGS[1014] = {[1]="Taupe Hamster";[2]="";}
-_PETSTRINGS[1015] = {[1]="Dusky Parrot";[2]="";}
-_PETSTRINGS[1016] = {[1]="Sunstone Drakeling";[2]="";}
-_PETSTRINGS[1017] = {[1]="Amethyst Drakeling";[2]=""}
-_PETSTRINGS[1018] = {[1]="Silver Scorpion";[2]=""}
-_PETSTRINGS[1019] = {[1]="Brown Oryx";[2]=""}
-_PETSTRINGS[1020] = {[1]="Armoured Scorpion";[2]=""}
-_PETSTRINGS[1021] = {[1]="Verdant Grodbog Queen";[2]=""}
-_PETSTRINGS[1022] = {[1]="Young Brown Oryx";[2]=""}
-_PETSTRINGS[1023] = {[1]="Terror of the Caverns";[2]=""}
-_PETSTRINGS[1024] = {[1]="Cobalt Grodbog Protector";[2]="";}
-_PETSTRINGS[1025] = {[1]="Obsidian Deep-claw Whelp";[2]="";}
-_PETSTRINGS[1026] = {[1]="Pony of Umbar";[2]="";}
-_PETSTRINGS[1027] = {[1]="Verdant Deep-claw Whelp";[2]="";}
-_PETSTRINGS[1028] = {[1]="Affable Grodbog Queen";[2]=""}
-_PETSTRINGS[1029] = {[1]="Poisonous Grodbog Protector";[2]=""}
-_PETSTRINGS[1030] = {[1]="Young Tawny Oryx";[2]=""}
-_PETSTRINGS[1031] = {[1]="Aggressive Grodbog Queen";[2]=""}
-_PETSTRINGS[1032] = {[1]="Lush Grodbog Protector";[2]=""}
-_PETSTRINGS[1033] = {[1]="Tame Hyena";[2]=""}
-_PETSTRINGS[1034] = {[1]="Blood Scorpion";[2]="";}
-_PETSTRINGS[1035] = {[1]="Pitch-dark Scorpion";[2]="";}
-_PETSTRINGS[1036] = {[1]="Pale Deep-claw Whelp";[2]="";}
-_PETSTRINGS[1037] = {[1]="Dusky Grodbog Queen";[2]=""}
-_PETSTRINGS[1038] = {[1]="Pale Scorpion";[2]=""}
-_PETSTRINGS[1039] = {[1]="Bright Grodbog Protector";[2]=""}
-_PETSTRINGS[1040] = {[1]="Striped Grodbog Protector";[2]=""}
-_PETSTRINGS[1041] = {[1]="Dark Grodbog Queen";[2]=""}
-_PETSTRINGS[1042] = {[1]="Copper Scorpion";[2]=""}
-_PETSTRINGS[1043] = {[1]="Tawny Oryx";[2]=""}
-_PETSTRINGS[1044] = {[1]="Striped Deep-claw Whelp";[2]="";}
-_PETSTRINGS[1045] = {[1]="Rusty Mûrai Herding Dog";[2]="";}
-_PETSTRINGS[1046] = {[1]="Toffee Mûrai Herding Dog";[2]="";}
-_PETSTRINGS[1047] = {[1]="Ginger Mûrai Herding Dog";[2]=""}
-_PETSTRINGS[1048] = {[1]="Desert Slime";[2]=""}
-_PETSTRINGS[1049] = {[1]="Armoured Desert Wasp";[2]=""}
-_PETSTRINGS[1050] = {[1]="Battle-worn Worm";[2]=""}
-_PETSTRINGS[1051] = {[1]="Brown Mûmak Calf";[2]=""}
-_PETSTRINGS[1052] = {[1]="Diligent Mûmak Calf";[2]=""}
-_PETSTRINGS[1053] = {[1]="Shade-born Zizanu";[2]=""}
-_PETSTRINGS[1054] = {[1]="Crafty Worm";[2]="";}
-_PETSTRINGS[1055] = {[1]="Industrious Mûmak Calf";[2]="";}
-_PETSTRINGS[1056] = {[1]="Mottled Vulture";[2]="";}
-_PETSTRINGS[1057] = {[1]="Marble Mûmak Calf";[2]=""}
-_PETSTRINGS[1058] = {[1]="Sun-laden Mûmak Calf";[2]=""}
-_PETSTRINGS[1059] = {[1]="Skulking Worm";[2]=""}
-_PETSTRINGS[1060] = {[1]="Sea-borne Slime";[2]=""}
-_PETSTRINGS[1061] = {[1]="Copper Zizanu";[2]=""}
-_PETSTRINGS[1062] = {[1]="Twilight Mûmak Calf";[2]=""}
-_PETSTRINGS[1063] = {[1]="Night-born Slime";[2]=""}
-_PETSTRINGS[1064] = {[1]="Rust Slime";[2]="";}
-_PETSTRINGS[1065] = {[1]="Burdened Mûmak Calf";[2]="";}
-_PETSTRINGS[1066] = {[1]="Marsh Crocodile Hatchling";[2]="";}
-_PETSTRINGS[1067] = {[1]="Stone Mûmak Calf";[2]=""}
-_PETSTRINGS[1068] = {[1]="Emerald Slime";[2]=""}
-_PETSTRINGS[1069] = {[1]="Overgrown Boglurker";[2]=""}
-_PETSTRINGS[1070] = {[1]="Ruby Worm";[2]=""}
-_PETSTRINGS[1071] = {[1]="Pale Zizanu";[2]=""}
-_PETSTRINGS[1072] = {[1]="Humming Sunbird";[2]=""}
-_PETSTRINGS[1073] = {[1]="Raspberry Slime";[2]=""}
+_PETSTRINGS[1004] = {[1]="Cream Corgi";[2]=""}
+_PETSTRINGS[1005] = {[1]="Library Corgi";[2]=""}
+_PETSTRINGS[1006] = {[1]="Winter's Wind";[2]=""}
+_PETSTRINGS[1007] = {[1]="Turquoise Snake";[2]=""}
+_PETSTRINGS[1008] = {[1]="Leafy Green Snake";[2]=""}
+_PETSTRINGS[1009] = {[1]="Fiery Desert Snake";[2]=""}
+_PETSTRINGS[1010] = {[1]="Chestnut Sand Worm";[2]=""}
+_PETSTRINGS[1011] = {[1]="Spotted Hamster";[2]=""}
+_PETSTRINGS[1012] = {[1]="Taupe Hamster";[2]=""}
+_PETSTRINGS[1013] = {[1]="Amethyst Drakeling";[2]=""}
+_PETSTRINGS[1014] = {[1]="Silver Scorpion";[2]="";}
+_PETSTRINGS[1015] = {[1]="Armoured Scorpion";[2]="";}
+_PETSTRINGS[1016] = {[1]="Verdant Grodbog Queen";[2]="";}
+_PETSTRINGS[1017] = {[1]="Terror of the Caverns";[2]=""}
+_PETSTRINGS[1018] = {[1]="Pony of Umbar";[2]=""}
+_PETSTRINGS[1019] = {[1]="Verdant Deep-claw Whelp";[2]=""}
+_PETSTRINGS[1020] = {[1]="Affable Grodbog Queen";[2]=""}
+_PETSTRINGS[1021] = {[1]="Poisonous Grodbog Protector";[2]=""}
+_PETSTRINGS[1022] = {[1]="Aggressive Grodbog Queen";[2]=""}
+_PETSTRINGS[1023] = {[1]="Lush Grodbog Protector";[2]=""}
+_PETSTRINGS[1024] = {[1]="Tame Hyena";[2]="";}
+_PETSTRINGS[1025] = {[1]="Blood Scorpion";[2]="";}
+_PETSTRINGS[1026] = {[1]="Pitch-dark Scorpion";[2]="";}
+_PETSTRINGS[1027] = {[1]="Pale Deep-claw Whelp";[2]="";}
+_PETSTRINGS[1028] = {[1]="Dusky Grodbog Queen";[2]=""}
+_PETSTRINGS[1029] = {[1]="Pale Scorpion";[2]=""}
+_PETSTRINGS[1030] = {[1]="Bright Grodbog Protector";[2]=""}
+_PETSTRINGS[1031] = {[1]="Striped Grodbog Protector";[2]=""}
+_PETSTRINGS[1032] = {[1]="Dark Grodbog Queen";[2]=""}
+_PETSTRINGS[1033] = {[1]="Striped Deep-claw Whelp";[2]=""}
+_PETSTRINGS[1034] = {[1]="Rusty Mûrai Herding Dog";[2]="";}
+_PETSTRINGS[1035] = {[1]="Toffee Mûrai Herding Dog";[2]="";}
+_PETSTRINGS[1036] = {[1]="Desert Slime";[2]="";}
+_PETSTRINGS[1037] = {[1]="Armoured Desert Wasp";[2]=""}
+_PETSTRINGS[1038] = {[1]="Battle-worn Worm";[2]=""}
+_PETSTRINGS[1039] = {[1]="Brown Mûmak Calf";[2]=""}
+_PETSTRINGS[1040] = {[1]="Diligent Mûmak Calf";[2]=""}
+_PETSTRINGS[1041] = {[1]="Shade-born Zizanu";[2]=""}
+_PETSTRINGS[1042] = {[1]="Crafty Worm";[2]=""}
+_PETSTRINGS[1043] = {[1]="Industrious Mûmak Calf";[2]=""}
+_PETSTRINGS[1044] = {[1]="Mottled Vulture";[2]="";}
+_PETSTRINGS[1045] = {[1]="Marble Mûmak Calf";[2]="";}
+_PETSTRINGS[1046] = {[1]="Sun-laden Mûmak Calf";[2]="";}
+_PETSTRINGS[1047] = {[1]="Skulking Worm";[2]=""}
+_PETSTRINGS[1048] = {[1]="Sea-borne Slime";[2]=""}
+_PETSTRINGS[1049] = {[1]="Copper Zizanu";[2]=""}
+_PETSTRINGS[1050] = {[1]="Twilight Mûmak Calf";[2]=""}
+_PETSTRINGS[1051] = {[1]="Night-born Slime";[2]=""}
+_PETSTRINGS[1052] = {[1]="Rust Slime";[2]=""}
+_PETSTRINGS[1053] = {[1]="Burdened Mûmak Calf";[2]=""}
+_PETSTRINGS[1054] = {[1]="Marsh Crocodile Hatchling";[2]="";}
+_PETSTRINGS[1055] = {[1]="Stone Mûmak Calf";[2]="";}
+_PETSTRINGS[1056] = {[1]="Emerald Slime";[2]="";}
+_PETSTRINGS[1057] = {[1]="Ruby Worm";[2]=""}
+_PETSTRINGS[1058] = {[1]="Pale Zizanu";[2]=""}
+_PETSTRINGS[1059] = {[1]="Raspberry Slime";[2]=""}