+++ Changelog.md
@@ -1,4 +1,9 @@
 Change log:
+* 2026-04-25 - U48 - 1 new + 14 newly available
+ * Quintessential Fawn
+ * Brown Jerboa, Dusky Parrot, Ginger Mûrai Herding Dog, Overgrown Boglurker, Sunstone Drakeling
+ * Brown Oryx, Young Brown Oryx, Humming Sunbird, Cobalt Grodbog Protector, Obsidian Deep-claw Whelp
+ * Tan Wolf-dog, Young Tawny Oryx, Copper Scorpion, Tawny Oryx
 * 2026-03-11 - U47.1 - 3 new
   * Marine Scorpion, Daffodil Celebration Kid, Tawny Aurochs
 * 2026-03-05 - U47 - 6 new + 26 future available